Revert "Merge branch 'trunk' into configurable-code-block-analysis"
This reverts commit aa868a2abc0dda2374a7ebc678c2b8954db5429f.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/configurable-code-block-analysis@39457 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cfa/ast/c/CInitializers.java
@@ -541,7 +541,7 @@ public final class CInitializers {
                 pInput -> {
                   CExpression index =
                       new CIntegerLiteralExpression(
-                          loc, CNumericTypes.INT, BigInteger.valueOf(pInput));
+                          loc, CNumericTypes.INT, BigInteger.valueOf(pInput.longValue()));
 
                   return new CArraySubscriptExpression(loc, elementType, currentSubobject, index);
                 })
